mam/v2/ntru: serialization and types - move
