+++ mam/v2/mam/BUILD
@@ -14,6 +14,8 @@ cc_library(
         "//mam/v2:defs",
         "//mam/v2/mss",
         "//mam/v2/ntru",
+        "//mam/v2/ntru:mam_ntru_pk_t_set",
+        "//mam/v2/ntru:mam_ntru_sk_t_set",
         "//mam/v2/pb3",
         "//mam/v2/prng",
     ],
@@ -64,8 +66,6 @@ cc_library(
     name = "mam_types",
     srcs = ["mam_types.c"],
     deps = [
-        ":mam_ntru_pk_t_set",
-        ":mam_ntru_sk_t_set",
         ":mam_pre_shared_key_t_set",
         ":mam_types_hdr",
         "//mam/v2:defs",
@@ -79,21 +79,5 @@ typed_container_generate(
     additional_include_path = "mam/v2/mam/mam_types.h",
     container_type = "set",
     parent_directory = "mam/v2/mam",
-    value_type = "mam_ntru_pk_t",
-)
-
-typed_container_generate(
-    additional_deps = ":mam_types_hdr",
-    additional_include_path = "mam/v2/mam/mam_types.h",
-    container_type = "set",
-    parent_directory = "mam/v2/mam",
     value_type = "mam_pre_shared_key_t",
-)
-
-typed_container_generate(
-    additional_deps = ":mam_types_hdr",
-    additional_include_path = "mam/v2/mam/mam_types.h",
-    container_type = "set",
-    parent_directory = "mam/v2/mam",
-    value_type = "mam_ntru_sk_t",
 )
