+++ mam/v2/mam/mam.h
@@ -16,10 +16,10 @@
 #include "mam/v2/mam/alloc.h"
 #include "mam/v2/mam/channel.h"
 #include "mam/v2/mam/endpoint.h"
-#include "mam/v2/mam/mam_ntru_pk_t_set.h"
 #include "mam/v2/mam/mam_pre_shared_key_t_set.h"
 #include "mam/v2/mam/mam_types.h"
 #include "mam/v2/mss/mss.h"
+#include "mam/v2/ntru/mam_ntru_pk_t_set.h"
 #include "mam/v2/ntru/ntru.h"
 #include "mam/v2/prng/prng.h"
 #include "mam/v2/sponge/sponge.h"
